Refactored test variables
