+++ tests/e2e/mode_config_build_test.py
@@ -9,10 +9,11 @@ from xmipp3_installer.installer.constants import paths
 from xmipp3_installer.shared import file_operations
 
 from . import get_cmake_project_path
-from .shell_command_outputs import mode_config_build
+from .shell_command_outputs import mode_cmake
+from .shell_command_outputs.mode_cmake import mode_config_build
 from .. import (
 	create_versions_json_file, get_assertion_message,
-	copy_file_from_reference, get_test_file
+	copy_file_from_reference
 )
 
 @pytest.mark.parametrize(
@@ -54,7 +55,7 @@ def __normalize_cmake_executable(raw_output: str) -> str: # CMake used deppends 
 	first_flag_index = raw_output.find(" -S")
 	text_up_to_cmake_exec = raw_output[:first_flag_index]
 	splitted_first_lines = text_up_to_cmake_exec.split("\n")
-	return "\n".join([splitted_first_lines[0], mode_config_build.CMAKE_EXECUTABLE]) + raw_output[first_flag_index:]
+	return "\n".join([splitted_first_lines[0], mode_cmake.CMAKE_EXECUTABLE]) + raw_output[first_flag_index:]
 
 def __normalize_generator_line(raw_output: str) -> str: # Generator used is sometimes printed, others not
 	return raw_output.replace(mode_config_build.GENERATOR_LINE, "")
@@ -73,12 +74,12 @@ def __normalize_paths(raw_output: str) -> str: # Absolute paths are different pe
 
 @pytest.fixture(params=[True])
 def __setup_evironment(request):
-	cmake_project_name = "valid" if request.param else "config_error"
+	cmake_project_name = mode_cmake.VALID_PROJECT if request.param else mode_cmake.CONFIG_ERROR_PROJECT
 	project_path = get_cmake_project_path(cmake_project_name)
 	try:
 		create_versions_json_file(output_path=project_path)
 		copy_file_from_reference(
-			get_test_file(os.path.join("conf-files", "input", "all-off.conf")),
+			mode_cmake.TEST_CONFIG_FILE_PATH,
 			os.path.join(project_path, paths.CONFIG_FILE)
 		)
 		yield get_cmake_project_path(cmake_project_name)
